requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
cryptocmd==0.6.0
cycler==0.10.0
Django==3.0.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.4
dnspython==2.1.0
idna==2.10
importlib-resources==5.1.3
Jinja2==3.0.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.0
matplotlib==3.4.2
multitasking==0.0.9
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
scipy==1.6.3
six==1.16.0
sqlparse==0.2.4
sweetviz==2.1.0
tablib==3.0.0
tqdm==4.60.0
urllib3==1.26.4
yfinance==0.1.59
